ChunkStorage without interior mutability
Removed a few locks, made things streaming when they should and
performance is way better, even if there is still a lot of bad and
inefficient code here and there.

Chunking and hash-tree node handling is very expensive, reducing
CHUNK_SIZE causes a *BIG* overhead. Even when everything is ironed out
and working fine, in real world use the size shouldn't be too small in
order to get the performance out of this transfer-chunks-thing.

Replaced some locks server-side with async versions (I'm not sure it's
the right thing, but for now it's one thing less to keep and eye on).